Android.bp: add headers-only export
Bionic is interested in using a subset of the functionality in highway.
Make that available.

Upstream support for headers-only builds is experimental, but has some
interest: https://github.com/google/highway/issues/2114. In Bionic, we
hope to use highway as a direct wrapper around simd intrinsics, rather
than using all of the fancier auto-dispatch logic.

Bug: 453997143
Test: mma in bionic
Change-Id: If0b1107b045d34636798ddd4d9d57f831514d633